Move Component SuccessOverlay to different folder
