+++ src/app/components/SuccessOverlay/SuccessPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import Button from '../../../components/Button/Button';
+import Button from '../Button/Button';
 
 export default function SuccessOverlay(): JSX.Element {
   return (
